move caching code to shared.go
